ML/requirements.txt
google-genai
pytesseract
easyocr
python-docx
spacy
nltk
pydantic
pydantic-settings
numpy
pandas
Pillow
pytest
streamlit
fastapi
uvicorn[standard]
datasets
transformers
torch
pyyaml
sentence-transformers
scikit-learn
faiss-cpu
python-multipart
